Create patch to remove problematic validation (#528)
Fixes https://github.com/pulumi/pulumi-azuredevops/issues/514

The validation always fails when running under the bridge because the
bridge fills in the default values for values which are marked as
ConflictsWith. Therefore, we'll remove this validation for the time
being to unblock users from using the variable group resource.

This can be removed once
https://github.com/pulumi/pulumi-terraform-bridge/issues/2618 is
resolved.